Energy and Inside Enemy
Regen rate is a bit slow on energy but it's a public variable so change
it however you want buddy boy.  Also you can now blink inside of an
enemy and you are stuck there.  I think I'll change it to make you a
child of the enemy but that sounds complex so for now it's this crude
way.